WORKAROUND: Use a singleton incognito context
Ideally, an incognito browser context should not be shared across
multiple RenderProcess instances. In the current state of things,
the RenderProcess created by blpwtk2 does not keep note on the
browser context it is associated with. This causes the browser
context manager to hand over a new context every time it is asked
to return the context associated with the RenderProcess and this
breaks many assumptions (since the context object keeps changing).
To workaround this problem, we treat the incognito context as
a singleton. This is not a problem for now since we don't mind
sharing the browser context across different embedders sharing
the same browser process.