Rework support for sideloaded fonts
With the removal of the renderer-side font cache, we no longer have a
way to alter the cache to insert side loaded fonts. Instead, we now
supply a custom font collection instead of the system default font
collection to the DirectWrite font proxy in the browser.

This custom font collection is a composite collection that use one or
more font collections. We supply this composite collection with the
system default font collection as well a private font collection. This
private font collection is an implementation of the DirectWrite font
collection interface that maps side loaded fonts to the process address
space and returns sections of it on demand.

It is important to note that the side loaded fonts are now loaded in
the browser process instead of the renderer process. This means that the
embedder must specify the side loaded fonts in the browser process
instead of the renderer process.