Revert "Remove extension when copying PDB files"
The original fix was meant to replace .dll.pdb and .exe.pdb to just
.pdb.  This causes a problem because the pdb filename encoded into the
actual module contains the .dll.pdb or .exe.pdb and this throws off the
debugger to look for a differently named .pdb file.

This commit is to revert the .pdb filenames back to the way it was. The
original change was made to workaround a problem with a pdb deployment tool.
This problem will be worked around some other way.