Eagerly load the UCRT's 'locale' API's upon 'DLL_PROCESS_ATTACH' (#109)
The 'DllMain()' for 'DLL_PROCESS_ATTACH' of each of blpwtk2's DLL's
calls UCRT's '__acrt_eagerly_load_locale_apis()'. This function calls
'GetProcAddress()' for each API needed for locales and caches the result
in a function table that exists per-process and per-DLL (because the UCRT
is statically linked).

This is necessary to avoid a later call by one thread to 'setlocale()'
initializating this function table in parallel to a new thread being attached
to the DLL. That will introduce a deadlock between the 'locale lock' and the
'loader lock': the first thread will have acquired the locale lock (because
the C++ streams library is initializing), and will be attempting to acquire
the loader lock (by calling 'GetProcAddress()'), while the second thread will
have acquired the loader lock (because the attachment work is done in
'_DllMainCRTStartup()') and will be attempting to acquire the locale lock
(to perform per-thread initialization).

This change should mediate this problem by forcing all DLL's to
'__acrt_eagerly_load_locale_apis()' to run once, per-process, by the process
that has already acquired the loader lock, before any attempts are made to
acquire the locale lock.

It is possible that loading the UCRT library dynamically instead of statically
could also correct this issue.