Always initialize the IPC channel in the child thread
Prior to this commit, the child process connected back to the browser only
if it had a channel name.  In the MojoIPC model, such channel name is not
needed and it's completely unnecessary.  The reason renderers were working
fine is because the channel name was set to a test value that was accidentally
checked-in in an earlier commit.  This was masking the real problem.

In this commit, all uses of the channel name in the child process is removed
and the child process connects back to the browser regardless the presence of
a channel name.