Do not enqueue 'WebPlugin' events if the document is inactive (#114)
The function 'blink::WebPluginContainerImpl::enqueueEvent()' is modified so
that it does nothing if its document is in an 'inactive' state in its lifecycle.
Such a state is reached before the document is initialized and while it is
'shutting down' (see the blink::DocumentLifecycle enum in DocumentLifecycle.h:
inactive states are 'Uninitialized', 'Inactive', 'Stopping', and 'Stopped').

A perhaps more thorough solution would be to notify the WebPlugin of changes
to the document's lifecycle, so that the plugin could batch up work to enqueue
once the document is active. In practice, though, the only 'inactive' state
where that would be useful is if the document is 'Uninitialized' or 'Inactive',
because it is not possible for the document to return from a 'Stopping' or
'Stopped' state to one in which it is active. And in an 'Unintialized' or
'Inactive' state, it is unlikely that the WebPlugin would exist in the first
place.