Fix collapsible spaces with word-break:break-all
Prior to this commit, the inline text iterator always rewinded to the last
mid-word break, even when it is touching a group of collapsible spaces.  This
broke the collapsing behavior of spaces in the beginning of the line when the
CSS property value word-break:break-all is used.

This commit conditionally calls rewindToFirstMidWordBreak() instead of
rewindToMidWordBreak() when the word-break mode is set to break-all and the
current iterator is touching a group of collapsible spaces from the beginning
of the line.